apps/backend/pyproject.toml
[project]
name = "webapp-api"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "supabase>=2.11.0",
    "pyyaml>=6.0",
    "boto3>=1.35.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.25.0",
]
